Initiated Setup for DisplayV2 and added 5780 proposal to documentation.
V2 of the display code will focus on making the timing more accurate. This will be done by having the display update at a fixed pace and having another core focus on making sure the desired RPM is maintained. There are many other deeper changes to the way the code works which is documented in a PDF that will be uploaded at a later date when we ensure the base implementation works.