Verified the timing requirements, built the code for showing full 3D frames with animations, optimized the datastorage method. Updated GPIO assignments, cleaned up code rot from I2C display, updated GPIO pin mapping to support the GPIO display.
Overall did a full update the code so we can show 3D animations. Fixed a lot of the code to look nice and did a number of optimizations. We verified that we can light the display fast enough and switch frames so that we can support 30 FPS. Next steps will be to get new data from blender and update our C# script to convert our blender data to our new data storage format. We will also start designing the new chassis as we have reduced the overall size needed for the design.